Add a bit of air control
