+++ Assets/Scripts/Player/State/PlayerFalling.cs
@@ -21,6 +21,7 @@ namespace Player.State
         public override void UpdatePhysics(Player player)
         {
             this.ApplyGravityBoost(player);
+            this.HandleHorizontalInput(player, boost: 0.15f);
             _reachedGround = player.OnGround();
         }
     }
