+++ Assets/Scripts/Player/State/PlayerState.cs
@@ -1,4 +1,6 @@
-﻿namespace Player.State
+﻿using UnityEngine;
+
+namespace Player.State
 {
     public abstract class PlayerState
     {
@@ -26,5 +28,34 @@
         protected void ApplyGravityBoost(Player player, float boost=1.6f) {
             player.Rigidbody.velocity += Vector2.up * Physics2D.gravity.y * boost * Time.deltaTime;
         }
+
+        protected void HandleHorizontalInput(Player player, float boost=1f) {
+            var horizontal = Input.GetAxis(Axis.Horizontal);
+
+            // If the player is changing direction  or hasn't reached MaxSpeedX yet
+            if (horizontal * player.Rigidbody.velocity.x < player.MaxSpeedX)
+            {
+                player.Rigidbody.AddForce(Vector2.right * horizontal * boost * player.MoveForceX);
+            }
+
+            if (Mathf.Abs(player.Rigidbody.velocity.x) > player.MaxSpeedX)
+            {
+                player.Rigidbody.velocity = new Vector2(Mathf.Sign(player.Rigidbody.velocity.x) * player.MaxSpeedX,
+                    player.Rigidbody.velocity.y);
+            }
+
+            if (horizontal < 0 && player.Facing == Facing.Right)
+            {
+                Utils.Flip(player.transform);
+                player.Facing = Facing.Left;
+            }
+            else if (horizontal > 0 && player.Facing == Facing.Left)
+            {
+                Utils.Flip(player.transform);
+                player.Facing = Facing.Right;
+            }
+
+        }
+
     }
 }
