+++ Assets/Scripts/Player/State/PlayerWalking.cs
@@ -30,30 +30,7 @@ namespace Player.State
 
         public override void UpdatePhysics(Player player)
         {
-            var horizontal = Input.GetAxis(Axis.Horizontal);
-
-            // If the player is changing direction  or hasn't reached MaxSpeedX yet
-            if (horizontal * player.Rigidbody.velocity.x < player.MaxSpeedX)
-            {
-                player.Rigidbody.AddForce(Vector2.right * horizontal * player.MoveForceX);
-            }
-
-            if (Mathf.Abs(player.Rigidbody.velocity.x) > player.MaxSpeedX)
-            {
-                player.Rigidbody.velocity = new Vector2(Mathf.Sign(player.Rigidbody.velocity.x) * player.MaxSpeedX,
-                    player.Rigidbody.velocity.y);
-            }
-
-            if (horizontal < 0 && player.Facing == Facing.Right)
-            {
-                Utils.Flip(player.transform);
-                player.Facing = Facing.Left;
-            }
-            else if (horizontal > 0 && player.Facing == Facing.Left)
-            {
-                Utils.Flip(player.transform);
-                player.Facing = Facing.Right;
-            }
+            this.HandleHorizontalInput(player);
         }
     }
 }
